[controller] final refactoring Signed-off-by: Viktor Kramarenko <viktor.kramarenko@flant.com>
