+++ images/sds-health-watcher-controller/pkg/controller/lvg_conditions_watcher_funcs.go
@@ -1,0 +1,68 @@
+package controller
+
+import (
+	"context"
+	"gopkg.in/yaml.v3"
+	v1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	"sds-health-watcher-controller/api/v1alpha1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+func getTargetConditionsCount(lvgCrd *v1.CustomResourceDefinition) (int, error) {
+	type item struct {
+		Type       string `json:"type"`
+		Properties struct {
+			LastTransitionTime struct {
+				Type string `json:"type"`
+			} `json:"lastTransitionTime"`
+			Message struct {
+				Type string `json:"type"`
+			} `json:"message"`
+			ObservedGeneration struct {
+				Type string `json:"type"`
+			} `json:"observedGeneration"`
+			Reason struct {
+				Type string `json:"type"`
+			} `json:"reason"`
+			Status struct {
+				Type string `json:"type"`
+			} `json:"status"`
+			Type struct {
+				Type string   `json:"type"`
+				Enum []string `json:"enum"`
+			} `json:"type"`
+		} `json:"properties"`
+	}
+	i := item{}
+
+	json, err := lvgCrd.Spec.Versions[0].Schema.OpenAPIV3Schema.Properties["status"].Properties["conditions"].Items.MarshalJSON()
+	if err != nil {
+		return 0, err
+	}
+
+	err = yaml.Unmarshal(json, &i)
+	if err != nil {
+		return 0, err
+	}
+
+	return len(i.Properties.Type.Enum), nil
+}
+
+func getLVGCRD(ctx context.Context, cl client.Client) (*v1.CustomResourceDefinition, error) {
+	crd := &v1.CustomResourceDefinition{}
+	err := cl.Get(ctx, client.ObjectKey{
+		Name: lvgCrdName,
+	}, crd)
+
+	return crd, err
+}
+
+func updateLVMVolumeGroupPhaseIfNeeded(ctx context.Context, cl client.Client, lvg *v1alpha1.LvmVolumeGroup, phase string) error {
+	if lvg.Status.Phase == phase {
+		return nil
+	}
+
+	lvg.Status.Phase = phase
+
+	return cl.Status().Update(ctx, lvg)
+}
